Sort the user accounts list when we're setting it.
Instead of trying to sort it during the task, we should sort it while
we're setting the value.  Ansible seems to have trouble taking the
output of the user_accounts.keys() after it's been stored in a variable.
Moving the sort to the time of the initial setting of the variable gets
us past the error.

This is the error that is resolved by this commit:

TASK [user_accounts : Add local user accounts to infrastructure system]
**************************************************************************************
fatal: [infra]: FAILED! => {"msg": "The task includes an option with an
undefined variable. The error was: 'dict object' has no attribute '
'\n\nThe error appears to be in
'/home/awalsh/bunsen/provisioning/roles/user_accounts/tasks/user_accounts.yml':
line 29, column 7, but may\nbe elsewhere in the file depending on the
exact syntax problem.\n\nThe offending line appears to be:\n\n    #
order, in order to ensure matching UID allocations!\n    - name: Add
local user accounts to infrastructure system\n      ^ here\n"}