Install Permatest from common package repo.
We no longer need to copy down all of the files to build it ourselves
via dkms and the various files that support it (like the modprobe.d
config) are all provided by the package.

We are continuing to include the step that loads the module, because it
won't automatically load until the system has been rebooted otherwise.