Removed reference to internal 'pbit-build' mount.
OSS Bunsen is intended to be 100% disconnected from the internal lab
environment.  The content that is provided by this reference is
not intended to be referenced in any way by the OSS bunsen.
(pbit-eng-common duplicates this code, which will be consolidated into
RPMs or into the VDO source tree itself)