Removed reference to internal 'file-server' repo
OSS Bunsen is intended to be 100% disconnected from the internal
environment.  The content that is provided by this reference is actually
also provided by the staged content that will otherwise be available via
github/COPR.

Additionally, one of the tasks creates a "dummy" directory, which is no
longer necessary since the package provides that instead.