Properly enable Permabit::RSVP module.
The previous spelling was overlooking the '  #config:' line in between
the 'Permabit::RSVP: ~' line and the rest of the configuration, so this
update properly addresses that issue.

Additionally, this change fixes up the comment about always using a
pristine copy of the configuration to be a declaration rather than
an #XXX: question.

Signed-off-by: Andrew Walsh <awalsh@redhat.com>