Configure Permabit::RSVP as needed in Bunsen.
Reworked the tweaking of /etc/permabit/perl.yaml to configure all of the
required parameters (defaultRSVPServer, emailDomain, jiraBrowseURL, and
jiraForceUser).  Note that three of these fields are actuall not
required for Permabit::RSVP to function outside of an assertion that
ensures they are configured, hence the 'fake' and 'not.required.com'
values.

Signed-off-by: Andrew Walsh <awalsh@redhat.com>