Added perl-Permabit-CurrentVersionFile.
This package is needed to be able to run 'make archive' in the linux-vdo
project, currently.  Spericifally it applies to the prepareGithubDist.pl
and prepareSourceDist.pl utilities.  However, there may be a more
general-purpose use for this package for versioning the various packages
we're maintaining out of the common project as well.  So rather than
moving it from one project to another, we split it out into a
sub-package to make sure it can be used where needed.

Signed-off-by: Andrew Walsh <awalsh@redhat.com>